NEW: Код разбит на модули. Настроено необходимое окружение
