+++ public/index.html
@@ -114,6 +114,6 @@
       </section>
     </main>
 
-    <script src="../src/main.js"></script>
+    <script src="bundle.js"></script>
   </body>
 </html>
